Restructure the BFT directory to follow new convention
